# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Collab API
  version: 0.1.0
  description: Local-first collaboration platform with RAG knowledge QA and AI task planning.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /admin/test-users:
    post:
      tags: [Admin]
      summary: Create a test user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTestUserRequest'
      responses:
        '201':
          description: Test user created
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
                  data:
                    type: object
                    properties:
                      userId: { type: string, format: uuid }
        '400':
          description: Request validation failed
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/users:
    get:
      tags: [Admin]
      summary: List local user accounts
      responses:
        '200':
          description: User list without password or token fields
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiMeta'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/AdminUserView' }
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /admin/users/{userId}/disable:
    parameters:
      - name: userId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Admin]
      summary: Disable a user account
      responses:
        '204':
          description: User disabled
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/users/{userId}/enable:
    parameters:
      - name: userId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Admin]
      summary: Enable a user account
      responses:
        '204':
          description: User enabled
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List current user's notifications
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { description: Notification list }
...
  /notifications/project/{projectId}:
    get:
      tags: [Notifications]
      summary: List current user's notifications in one project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200': { description: Project notification list }
        '404': { $ref: '#/components/responses/NotFound' }
  /notifications/{notificationId}/read:
    patch:
      tags: [Notifications]
      summary: Mark one owned notification as read
      parameters:
        - { name: notificationId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Marked as read }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Forbidden:
      description: Forbidden
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Conflict:
      description: Conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  schemas:
    ApiMeta:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiMeta'
        - type: object
          required: [data]
          properties:
            data: { type: 'null' }
...
    CreateTestUserRequest:
      type: object
      required: [username, displayName, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 40 }
        displayName: { type: string, minLength: 1, maxLength: 60 }
        password: { type: string, minLength: 6, maxLength: 100 }
    AdminUserView:
      type: object
      required: [id, username, displayName, status, systemAdmin, createdAt]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        email: { type: [string, 'null'], format: email }
        status: { type: string, enum: [ACTIVE, DISABLED] }
        systemAdmin: { type: boolean }
        lastLoginAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
...